Updated test data for FDP-501 - changed feature and step definition file, request and response POJO files and Cart_TestData.java file

+++ src/test/java/model/Response/Cart/GetCartRes/CoPay.java
@@ -5,15 +5,15 @@ import java.util.List;
 
 public class CoPay {
 
-    private Integer percentage;
+    private Double percentage;
     private List<Integer> premium = null;
     private List<Integer> taxAndServicePremium = null;
 
-    public Integer getPercentage() {
+    public Double getPercentage() {
         return percentage;
     }
 
-    public void setPercentage(Integer percentage) {
+    public void setPercentage(Double percentage) {
         this.percentage = percentage;
     }
